+++ src/framework/multiwindows/internal/multiprocess/multiprocessprovider.h
@@ -61,7 +61,7 @@ public:
     void init();
 
     // Contexts info
-    int windowCount() const override;
+    size_t windowCount() const override;
     bool isFirstWindow() const override;
 
     // Project opening
